# fix: Prevent excessively copying RawCacheKey