--- a/crates/cache/src/lib.rs
+++ b/crates/cache/src/lib.rs
@@ -123,8 +123,15 @@ impl CacheKey<'_> {
     }
 }
 
-#[derive(Hash, Clone, Copy, Eq, PartialEq)]
+#[derive(Hash, Eq, PartialEq)]
 pub struct RawCacheKey(u64);
+
+impl RawCacheKey {
+    #[must_use]
+    pub fn as_u64(&self) -> u64 {
+        self.0
+    }
+}
 
 impl QueryResult {
     #[must_use]
@@ -149,9 +156,9 @@ pub struct CachedQueryResult {
 #[async_trait]
 pub trait QueryResultCache {
     async fn get<'a>(&self, key: CacheKey<'a>) -> Result<Option<CachedQueryResult>>;
-    async fn get_raw_key(&self, raw_key: RawCacheKey) -> Result<Option<CachedQueryResult>>;
+    async fn get_raw_key(&self, raw_key: &RawCacheKey) -> Result<Option<CachedQueryResult>>;
     async fn put<'a>(&self, key: CacheKey<'a>, result: CachedQueryResult) -> Result<()>;
-    async fn put_raw_key(&self, raw_key: RawCacheKey, result: CachedQueryResult) -> Result<()>;
+    async fn put_raw_key(&self, raw_key: &RawCacheKey, result: CachedQueryResult) -> Result<()>;
     async fn invalidate_for_table(&self, table_name: TableReference) -> Result<()>;
     fn size_bytes(&self) -> u64;
     fn item_count(&self) -> u64;
@@ -215,13 +222,13 @@ impl QueryResultsCacheProvider {
     /// Will return `Err` if method fails to access the cache
     pub async fn get(&self, key: CacheKey<'_>) -> Result<Option<CachedQueryResult>> {
         let raw_key = key.as_raw_key();
-        self.get_raw_key(raw_key).await
+        self.get_raw_key(&raw_key).await
     }
 
     /// # Errors
     ///
     /// Will return `Err` if method fails to access the cache
-    pub async fn get_raw_key(&self, raw_key: RawCacheKey) -> Result<Option<CachedQueryResult>> {
+    pub async fn get_raw_key(&self, raw_key: &RawCacheKey) -> Result<Option<CachedQueryResult>> {
         metrics::REQUESTS.add(1, &[]);
         match self.cache.get_raw_key(raw_key).await {
             Ok(Some(cached_result)) => {
@@ -245,7 +252,11 @@ impl QueryResultsCacheProvider {
     /// # Errors
     ///
     /// Will return `Err` if method fails to access the cache
-    pub async fn put_raw_key(&self, raw_key: RawCacheKey, result: CachedQueryResult) -> Result<()> {
+    pub async fn put_raw_key(
+        &self,
+        raw_key: &RawCacheKey,
+        result: CachedQueryResult,
+    ) -> Result<()> {
         let res = self.cache.put_raw_key(raw_key, result).await;
         self.report_size_metrics();
         res