## 📝 Summary

<!-- What does this PR change? Why is it necessary? Keep it concise. -->

1. Remove the `Clone` and `Copy` implementations from `RawCacheKey`. This is a newtype that wraps a `u64` - the `u64` can be accessed directly through a borrow on `&self`, as `u64` itself implements `Copy`. This prevents a situation where a `RawCacheKey` is copied, only for the underlying `u64` to also be copied out of the newtype (e.g. when the newtype is hashed, etc).
    * Also includes associated updates to trait bounds for `QueryResultCache` and `QueryResultsCacheProvider` to accept borrowed values of `RawCacheKey` instead of owned values.
2. Prevent calculating the `RawCacheKey` inside `get_or_create_logical_plan` - if a cache miss occurs, the `RawCacheKey` will already be computed, so there is no reason to re-compute.
3. Updates the logical plan cache to use keys of `u64`, instead of keys of `RawCacheKey`. This aligns with existing behavior for the `LruCache`, which uses the `u64` value from the `RawCacheKey`.

Performing testing with `oha`, some headline metrics are observed:

* Before:
    * `cache_key_type: plan` - 153,438 req/s, P99 @ 0.0011
    * `cache_key_type: sql` - 160,441 req/s, P99 @ 0.0010
* After:
    * `cache_key_type: plan` - 184,764 req/s, P99 @ 0.0009
    * `cache_key_type: sql` - 199,520 req/s, P99 @ 0.0009

Translates to a roughly 20% improvement in query throughput, and reduction in P99 query latency.

## 🔗 Related

<!-- Link to relevant issues, discussions, or other PRs. Use "Closes #123" to auto-close issues. Omit if none. -->

* Part of #5832